requirements.txt
nextcord==2.5.0
python-dotenv>=1.0.0
psycopg2>=2.9.6
apscheduler>=3.10.1
